REFACTOR: per suggestions in pull request.
