+++ perma_web/perma/settings/utils/post_processing.py
@@ -33,6 +33,10 @@ def post_process_settings(settings):
             'task': 'perma.tasks.upload_all_to_internet_archive',
             'schedule': crontab(minute='0', hour='*'),
         },
+        'cm-sync': {
+            'task': 'perma.tasks.cm_sync',
+            'schedule': crontab(hour='3'),
+        },
 
     }
     settings['CELERYBEAT_SCHEDULE'] = dict(((job, celerybeat_job_options[job]) for job in settings.get('CELERYBEAT_JOB_NAMES', [])),
